requirements.txt
astunparse==1.6.3
cycler==0.12.1
dnspython==2.7.0
docker-pycreds==0.4.0
expecttest==0.2.1
ffmpeg-python==0.2.0
fonttools==4.55.0
fsspec==2024.10.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
hypothesis==6.115.5
kiwisolver==1.4.7
lintrunner==0.12.5
llvmlite==0.43.0
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.1
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python-headless==4.10.0.84
optree==0.13.0
pandas==2.2.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
protobuf==5.28.3
python-dateutil==2.9.0.post0
scikit-learn==1.5.2
scipy==1.14.1
sentry-sdk==2.18.0
setproctitle==1.3.4
smmap==5.0.1
sortedcontainers==2.4.0
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchelastic==0.2.2
torchvision==0.20.1+cu124
triton==3.1.0
types-dataclasses==0.6.6
wandb==0.18.7
zstandard==0.23.0
